Unify MCAP parsing into `re_mcap`
