+++ ARCHITECTURE.md
@@ -98,11 +98,11 @@ Of course, this will only take us so far. In the future we plan on caching queri
 Here is an overview of the crates included in the project:
 
 <picture>
-  <img src="https://static.rerun.io/crates/985f26b3b7cf2cac523b004de742356e3f263da7/full.png" alt="">
-  <source media="(max-width: 480px)" srcset="https://static.rerun.io/crates/985f26b3b7cf2cac523b004de742356e3f263da7/480w.png">
-  <source media="(max-width: 768px)" srcset="https://static.rerun.io/crates/985f26b3b7cf2cac523b004de742356e3f263da7/768w.png">
-  <source media="(max-width: 1024px)" srcset="https://static.rerun.io/crates/985f26b3b7cf2cac523b004de742356e3f263da7/1024w.png">
-  <source media="(max-width: 1200px)" srcset="https://static.rerun.io/crates/985f26b3b7cf2cac523b004de742356e3f263da7/1200w.png">
+  <img src="https://static.rerun.io/crates/acbc767554b61f925c6aad1356adf2e7e51692bd/full.png" alt="An overview of crates in the Rerun ecosystem.">
+  <source media="(max-width: 480px)" srcset="https://static.rerun.io/crates/acbc767554b61f925c6aad1356adf2e7e51692bd/480w.png">
+  <source media="(max-width: 768px)" srcset="https://static.rerun.io/crates/acbc767554b61f925c6aad1356adf2e7e51692bd/768w.png">
+  <source media="(max-width: 1024px)" srcset="https://static.rerun.io/crates/acbc767554b61f925c6aad1356adf2e7e51692bd/1024w.png">
+  <source media="(max-width: 1200px)" srcset="https://static.rerun.io/crates/acbc767554b61f925c6aad1356adf2e7e51692bd/1200w.png">
 </picture>
 
 <!-- !!! IMPORTANT!!!
@@ -227,7 +227,7 @@ Update instructions:
 | re_format          | Miscellaneous tools to format and parse numbers, durations, etc.                     |
 | re_int_histogram   | A histogram with `i64` keys and `u32` counts, supporting both sparse and dense uses. |
 | re_log             | Helpers for setting up and doing text logging in the Rerun crates.                   |
-| re_mcap_ros2       | Deserialization of a subset of ROS2 messages.                                        |
+| re_mcap            | Deserialization and helpers for MCAP files.                                          |
 | re_memory          | Run-time memory tracking and profiling.                                              |
 | re_perf_telemetry  | In and out of process performance profiling utilities for Rerun & Redap              |
 | re_smart_channel   | A channel that keeps track of latency and queue length.                              |
