+++ crates/store/re_data_loader/Cargo.toml
@@ -31,7 +31,7 @@ re_error.workspace = true
 re_log.workspace = true
 re_log_encoding = { workspace = true, features = ["decoder"] }
 re_log_types.workspace = true
-re_mcap_ros2.workspace = true
+re_mcap.workspace = true
 re_smart_channel.workspace = true
 re_tracing.workspace = true
 re_types = { workspace = true, features = ["ecolor", "glam", "image", "video"] }
@@ -41,7 +41,6 @@ anyhow.workspace = true
 arrow.workspace = true
 byteorder.workspace = true
 memmap2.workspace = true
-cdr-encoding.workspace = true
 crossbeam.workspace = true
 image.workspace = true
 indexmap.workspace = true
@@ -49,7 +48,6 @@ itertools.workspace = true
 notify.workspace = true
 mcap.workspace = true
 parking_lot.workspace = true
-prost-reflect.workspace = true
 rayon.workspace = true
 serde.workspace = true
 serde_json.workspace = true
