--- crates/store/re_data_loader/src/mcap/mod.rs
@@ -1,10 +1,0 @@
-//! Library providing utilities to load MCAP files with Rerun.
-
-pub mod cdr;
-pub(crate) mod dds;
-pub mod decode;
-pub mod layers;
-pub mod schema;
-pub mod util;
-
-pub use layers::{Layer, MessageLayer};
