+++ crates/utils/re_mcap/src/lib.rs
@@ -1,0 +1,10 @@
+//! Library providing utilities to load MCAP files with Rerun.
+
+mod layers;
+
+pub(crate) mod parsers;
+pub(crate) mod util;
+
+pub use layers::{Layer, LayerIdentifier, LayerRegistry, MessageLayer, SelectedLayers};
+// TODO(grtlr): We should expose an `Mcap` object that internally holds the summary + a reference to the bytes.
+pub use util::read_summary;
