+++ crates/utils/re_mcap/src/parsers/ros2msg/definitions/builtin_interfaces.rs
