+++ crates/utils/re_mcap/src/parsers/ros2msg/definitions/geometry_msgs.rs
