+++ crates/utils/re_mcap/src/parsers/ros2msg/definitions/sensor_msgs.rs
@@ -6,7 +6,7 @@ use std::borrow::Cow;
 
 use serde::{Deserialize, Serialize};
 
-use crate::{geometry_msgs, std_msgs::Header};
+use super::{geometry_msgs, std_msgs::Header};
 
 /// This message contains an uncompressed image.
 ///
