+++ crates/utils/re_mcap/src/parsers/ros2msg/mod.rs
@@ -1,0 +1,12 @@
+mod definitions;
+
+pub mod sensor_msgs;
+pub mod std_msgs;
+
+use std::sync::Arc;
+
+use arrow::{
+    array::{FixedSizeListBuilder, ListBuilder, UInt8Builder},
+    datatypes::{DataType, Field},
+};
+use re_types::{Loggable as _, components};
