+++ crates/utils/re_mcap/src/parsers/ros2msg/sensor_msgs/compressed_image.rs
@@ -1,16 +1,16 @@
+use super::super::definitions::sensor_msgs;
 use re_chunk::{
     Chunk, ChunkId, RowId, TimePoint,
     external::arrow::array::{FixedSizeListBuilder, StringBuilder},
 };
 use re_log_types::TimeCell;
-use re_mcap_ros2::sensor_msgs;
 use re_types::{
     ComponentDescriptor,
     archetypes::{EncodedImage, VideoStream},
     components::VideoCodec,
 };
 
-use crate::mcap::{
+use crate::parsers::{
     cdr,
     decode::{McapMessageParser, ParserContext, PluginError},
 };
