--- crates/utils/re_mcap_ros2/Cargo.toml
@@ -1,20 +1,0 @@
-[package]
-name = "re_mcap_ros2"
-description = "Deserialization of a subset of ROS2 messages."
-authors.workspace = true
-edition.workspace = true
-homepage.workspace = true
-include.workspace = true
-license.workspace = true
-publish = true
-repository.workspace = true
-rust-version.workspace = true
-version.workspace = true
-
-[lints]
-workspace = true
-
-
-[dependencies]
-serde.workspace = true
-serde_bytes.workspace = true
