--- crates/utils/re_mcap_ros2/README.md
@@ -1,10 +1,0 @@
-# re_mcap_ros2
-
-Part of the [`rerun`](https://github.com/rerun-io/rerun) family of crates.
-
-[![Latest version](https://img.shields.io/crates/v/re_mcap_ros2)](https://crates.io/crates/re_mcap_ros2?speculative-link)
-[![Documentation](https://docs.rs/re_mcap_ros2/badge.svg)](https://docs.rs/re_mcap_ros2?speculative-link)
-![MIT](https://img.shields.io/badge/license-MIT-blue.svg)
-![Apache](https://img.shields.io/badge/license-Apache-blue.svg)
-
-Deserialization helpers for a subset of ROS2 messages.
